stock-prediction/requirements.txt
tensorflow
tensorflow_hub
bert-for-tf2
#pytorch
numpy
pandas
matplotlib
nltk
yfinance
news-please
google
scikit-learn